requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1
Flask==3.0.2
rich==13.7.1
pytest==8.1.1
pytest-mock==3.14.0
